replace create appointments component
