+++ src/pages/CreateCourse.tsx
@@ -8,9 +8,9 @@ import { LFSubject } from '../types/lernfair/Subject';
 
 import InstructionProgress from '../widgets/InstructionProgress';
 
-import CourseAppointments from './course-creation/CourseAppointments';
 import CourseData from './course-creation/CourseData';
 import CoursePreview from './course-creation/CoursePreview';
+import CourseAppointments from './course-creation/CourseAppointments';
 
 import { DateTime } from 'luxon';
 import { LFInstructor, LFLecture, LFSubCourse, LFTag } from '../types/lernfair/Course';
@@ -26,7 +26,6 @@ import { GraphQLError } from 'graphql';
 import AsNavigationItem from '../components/AsNavigationItem';
 import { BACKEND_URL } from '../config';
 import NotificationAlert from '../components/notifications/NotificationAlert';
-import CreateCourseAppointments from './course-creation/CreateCourseAppointments';
 
 export type CreateCourseError = 'course' | 'subcourse' | 'set_image' | 'upload_image' | 'instructors' | 'lectures' | 'tags';
 
@@ -866,8 +865,7 @@ const CreateCourse: React.FC = () => {
                                     onRemoveInstructor={removeInstructor}
                                 />
                             )}
-                            {/* {currentIndex === 1 && <CourseAppointments onNext={onNext} onBack={onBack} onDeleteAppointment={deleteAppointment} />} */}
-                            {currentIndex === 1 && <CreateCourseAppointments next={onNext} back={onBack} />}
+                            {currentIndex === 1 && <CourseAppointments next={onNext} back={onBack} />}
                             {currentIndex === 2 && (
                                 <>
                                     <CoursePreview
